Ensure min value of max_recording_time_limit
Problem statement:
1. More than once we have had to revert a manual increase to the system_config maximum call recording setting.  I believe that there is a feature to sync the setting to the schema value and that might be related.
2. It's required that the system-config default should be 10800, as seconds in 3 hours.

Since I found no evidence of auto-sync in code, and provision of max_recording_time_limit is expected using `sup kapps_config set_integer "media" "max_recording_time_limit" 3601`, there is almost no trail of manual updated such attribute. So I decide to enforce a mininum value of this attribute even when it can be set to a lower time limit.

The mininum value is purposely set to the default minus 1 second. It is easier to tell whether the value has been manually set lower then the threshold.

We could introduce a threshold or ceiling/floor for an attribute. But I think this would be an over haul in the code. That may not be warranted to be implemented at the moment.