Set correct `used_by` in number doc when port_in is completed
An earlier bug allows `used_by` is not set when a port_in number is
in `rejected` state. Although this has been fix in this PR,
the ramification of earlier numbers without this fix still exists.
Therefore, this commit checks if `used_by` is not set properly when
the port_in transitions to `completed` state.

It also fixes another bug that created phone_number as a result of
port_in completed will preserve `used_by` app. That has been missing.